+++ cmd/kubelet/app/BUILD
@@ -71,32 +71,7 @@ go_library(
         "//pkg/util/oom:go_default_library",
         "//pkg/util/rlimit:go_default_library",
         "//pkg/version:go_default_library",
-        "//pkg/volume:go_default_library",
-        "//pkg/volume/aws_ebs:go_default_library",
-        "//pkg/volume/azure_dd:go_default_library",
-        "//pkg/volume/azure_file:go_default_library",
-        "//pkg/volume/cephfs:go_default_library",
-        "//pkg/volume/cinder:go_default_library",
-        "//pkg/volume/configmap:go_default_library",
-        "//pkg/volume/downwardapi:go_default_library",
-        "//pkg/volume/empty_dir:go_default_library",
-        "//pkg/volume/fc:go_default_library",
-        "//pkg/volume/flexvolume:go_default_library",
-        "//pkg/volume/flocker:go_default_library",
-        "//pkg/volume/gce_pd:go_default_library",
-        "//pkg/volume/git_repo:go_default_library",
-        "//pkg/volume/glusterfs:go_default_library",
-        "//pkg/volume/host_path:go_default_library",
-        "//pkg/volume/iscsi:go_default_library",
-        "//pkg/volume/nfs:go_default_library",
-        "//pkg/volume/photon_pd:go_default_library",
-        "//pkg/volume/portworx:go_default_library",
-        "//pkg/volume/projected:go_default_library",
-        "//pkg/volume/quobyte:go_default_library",
-        "//pkg/volume/rbd:go_default_library",
-        "//pkg/volume/scaleio:go_default_library",
-        "//pkg/volume/secret:go_default_library",
-        "//pkg/volume/vsphere_volume:go_default_library",
+        "//pkg/volume/probeplugins:go_default_library",
         "//vendor:github.com/golang/glog",
         "//vendor:github.com/spf13/cobra",
         "//vendor:github.com/spf13/pflag",
