+++ pkg/api/validation/BUILD
@@ -15,7 +15,6 @@ go_library(
         "events.go",
         "schema.go",
         "validation.go",
-        "volume_plugins.go",
     ],
     tags = ["automanaged"],
     deps = [
@@ -27,30 +26,6 @@ go_library(
         "//pkg/capabilities:go_default_library",
         "//pkg/features:go_default_library",
         "//pkg/security/apparmor:go_default_library",
-        "//pkg/volume:go_default_library",
-        "//pkg/volume/aws_ebs:go_default_library",
-        "//pkg/volume/azure_dd:go_default_library",
-        "//pkg/volume/azure_file:go_default_library",
-        "//pkg/volume/cephfs:go_default_library",
-        "//pkg/volume/cinder:go_default_library",
-        "//pkg/volume/configmap:go_default_library",
-        "//pkg/volume/downwardapi:go_default_library",
-        "//pkg/volume/empty_dir:go_default_library",
-        "//pkg/volume/fc:go_default_library",
-        "//pkg/volume/flexvolume:go_default_library",
-        "//pkg/volume/flocker:go_default_library",
-        "//pkg/volume/gce_pd:go_default_library",
-        "//pkg/volume/git_repo:go_default_library",
-        "//pkg/volume/glusterfs:go_default_library",
-        "//pkg/volume/host_path:go_default_library",
-        "//pkg/volume/iscsi:go_default_library",
-        "//pkg/volume/nfs:go_default_library",
-        "//pkg/volume/photon_pd:go_default_library",
-        "//pkg/volume/projected:go_default_library",
-        "//pkg/volume/quobyte:go_default_library",
-        "//pkg/volume/rbd:go_default_library",
-        "//pkg/volume/secret:go_default_library",
-        "//pkg/volume/vsphere_volume:go_default_library",
         "//vendor:github.com/emicklei/go-restful/swagger",
         "//vendor:github.com/exponent-io/jsonpath",
         "//vendor:github.com/golang/glog",
@@ -100,7 +75,6 @@ go_test(
         "//pkg/apis/extensions/v1beta1:go_default_library",
         "//pkg/capabilities:go_default_library",
         "//pkg/security/apparmor:go_default_library",
-        "//pkg/volume:go_default_library",
         "//vendor:github.com/ghodss/yaml",
         "//vendor:k8s.io/apimachinery/pkg/api/resource",
         "//vendor:k8s.io/apimachinery/pkg/api/testing",
