+++ pkg/volume/BUILD
@@ -112,6 +112,7 @@ filegroup(
         "//pkg/volume/nfs:all-srcs",
         "//pkg/volume/photon_pd:all-srcs",
         "//pkg/volume/portworx:all-srcs",
+        "//pkg/volume/probeplugins:all-srcs",
         "//pkg/volume/projected:all-srcs",
         "//pkg/volume/quobyte:all-srcs",
         "//pkg/volume/rbd:all-srcs",
@@ -119,6 +120,7 @@ filegroup(
         "//pkg/volume/secret:all-srcs",
         "//pkg/volume/testing:all-srcs",
         "//pkg/volume/util:all-srcs",
+        "//pkg/volume/validation:all-srcs",
         "//pkg/volume/vsphere_volume:all-srcs",
     ],
     tags = ["automanaged"],
