+++ pkg/volume/validation/pv_validation.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2016 The Kubernetes Authors.
+Copyright 2017 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -17,15 +17,18 @@ limitations under the License.
 package validation
 
 import (
+	"github.com/golang/glog"
 	"k8s.io/apimachinery/pkg/util/validation/field"
 	"k8s.io/kubernetes/pkg/api"
+	"k8s.io/kubernetes/pkg/api/v1"
 	"k8s.io/kubernetes/pkg/volume"
+	"k8s.io/kubernetes/pkg/volume/probeplugins"
 )
 
 var volumePlugins []volume.VolumePlugin
 
 func init() {
-	volumePlugins = probeVolumePlugins()
+	volumePlugins = probeplugins.ProbeVolumePlugins("")
 }
 
 // ValidateForMountOptions validations mount options
@@ -45,4 +48,32 @@ func ValidateForMountOptions(pv *api.PersistentVolume) field.ErrorList {
 		}
 	}
 	return allErrs
+}
+
+func findPluginBySpec(volumePlugins []volume.VolumePlugin, pv *api.PersistentVolume) volume.VolumePlugin {
+	matches := []volume.VolumePlugin{}
+	v1Pv := &v1.PersistentVolume{}
+	err := v1.Convert_api_PersistentVolume_To_v1_PersistentVolume(pv, v1Pv, nil)
+	if err != nil {
+		glog.Errorf("Error converting to v1.PersistentVolume: %v", err)
+		return nil
+	}
+	volumeSpec := &volume.Spec{PersistentVolume: v1Pv}
+	for _, plugin := range volumePlugins {
+		if plugin.CanSupport(volumeSpec) {
+			matches = append(matches, plugin)
+		}
+	}
+
+	if len(matches) == 0 {
+		glog.V(5).Infof("No matching plugin found for : %s", pv.Name)
+		return nil
+	}
+
+	if len(matches) > 1 {
+		glog.V(3).Infof("multiple volume plugins matched for : %s ", pv.Name)
+		return nil
+	}
+
+	return matches[0]
 }
